yolo-service/requirements.txt
ultralytics==8.0.196
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.4
requests==2.31.0